Impl adapter
